luna/requirements.txt
# Luna Discord Bot Requirements
# Core dependencies
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0

# AI integration
google-generativeai>=0.3.0

# Optional for development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
